Added NIF values for overseas fencers
I've added a section that calculates the NIF total of all of the fencers below a specific fencer. If the fencer is not from the UK AND they do not have a valid BFA_ID according to the rankings, then they are "overseas". This means that their NIF values is calculated based on the number of people they are better than.