Added extra check for Int fencers
Added an extra check to make sure that the LastName matches for international fencers. This is to stop duplicates, or more than one possible identity when there are two people with the same position. This shouldn't have any affect on the NIF value, but it is better to be safe than sorry!