Added International Results
Added a section to check whether a fencer is on the internation rankings list. This is to determine whether they are an "overseas" fencer.

Added a comparision between overseas fencer rankings due to their world rankings and their NIF due to beating naitonal fencers. The larger between these two NIF values is chosen as the "notional" NIF of the fencer.